+++ .gitignore
@@ -1,2 +1,4 @@
 node_modules/
 .Ds_Store
+script.txt
+ex_modules
