+++ src/index.html
@@ -14,7 +14,7 @@
   <section class="container">
     <div class="header">
       <h1>Today's To Do </h1>
-      <i style="font-size:24px" class="fa">&#xf021;</i>
+      <i class="fa refresh">&#xf021;</i>
     </div>
     <form class="form">
       <input type="text" class="add" placeholder="Add to your list...">
