+++ src/style.css
@@ -26,6 +26,10 @@ body {
   color: #787878;
 }
 
+.refresh {
+  font-size: 24px;
+}
+
 h1 {
   text-align: center;
   margin-left: 20px;
@@ -68,7 +72,6 @@ h1 {
   padding: 20px;
   color: #787878;
 }
-
 
 .strike {
   color: rgb(113, 109, 109);
